[TASK] Convert the `DefaultController` unit tests to functionals
This class has dependencies and will soon use DI. So will no be
able to keep using unit tests for it.
