+++ Tests/Functional/FrontEnd/DefaultController/ListViewTest.php
@@ -62,6 +62,53 @@ final class ListViewTest extends FunctionalTestCase
     /**
      * @test
      */
+    public function eventListFlavorWithoutUidCreatesListView(): void
+    {
+        $controller = $this->createPartialMock(
+            TestingDefaultController::class,
+            [
+                'createListView',
+                'createSingleView',
+                'pi_initPIflexForm',
+                'getTemplateCode',
+                'setLabels',
+                'createHelperObjects',
+            ],
+        );
+        $controller->expects(self::once())->method('createListView')->with('seminar_list');
+        $controller->expects(self::never())->method('createSingleView');
+
+        $controller->piVars = [];
+
+        $controller->main('', ['what_to_display' => 'seminar_list']);
+    }
+
+    /**
+     * @test
+     */
+    public function eventListFlavorWithUidCreatesListView(): void
+    {
+        $controller = $this->createPartialMock(
+            TestingDefaultController::class,
+            [
+                'createListView',
+                'createSingleView',
+                'pi_initPIflexForm',
+                'getTemplateCode',
+                'setLabels',
+                'createHelperObjects',
+            ],
+        );
+        $controller->expects(self::once())->method('createListView')->with('seminar_list');
+        $controller->expects(self::never())->method('createSingleView');
+
+        $controller->piVars = ['showUid' => '42'];
+
+        $controller->main('', ['what_to_display' => 'seminar_list']);
+    }
+    /**
+     * @test
+     */
     public function listViewShowsVisibleSingleEvent(): void
     {
         $subject = $this->buildSubjectForListView('EventList');
